chi haja

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -8,9 +8,9 @@
 
 int	_atoi(char *s)
 {
-	int		i;
-	long	res;
-	int		sign;
+	int	i;
+	int	res;
+	int	sign;
 
 	i = 0;
 	res = 0;
@@ -23,5 +23,5 @@ int	_atoi(char *s)
 	}
 	while (s[i] >= '0' && s[i] <= '9')
 		res = res * 10 + s[i++] - '0';
-	return ((int)res * sign);
+	return (res * sign);
 }
